rt-detr/requirements.txt
# PyTorch con CUDA 11.8
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0

# YOLO y Computer Vision (Ultralytics includes RT-DETR)
ultralytics
opencv-python-headless

# Transformers y DINO
transformers
supervision
groundingdino-py

# Visualización y utilidades
tensorboard
matplotlib
pillow
pyyaml
tqdm

# Hyperparameter optimization
optuna
joblib

#roboflow
roboflow==1.2.13
python-dotenv==1.2.1

# NumPy 1.26.4 (última 1.x estable compatible con PyTorch 2.1)
numpy==1.26.4
